+++ src/idc/imgaug/filter/__init__.py
@@ -1,3 +1,4 @@
+from ._aruco_detector import ArucoDetector, DEFAULT_ARUCO_TYPE, ARUCO_TYPES
 from ._augment_util import augment_image
 from ._base_filter import BaseFilter
 from ._base_image_augmentation import BaseImageAugmentation